The new backend to compare tree works perfectly.
The front end has been heavily changed and, by using the libraty bigtree, it's order of magnitudes faster. This is also due to the massive use I am doing of async coros.

Despite all of these changes, the front-end works now as before (or event better in terms of bugs and responsiveness).

Job to be done for the next time: re-code all of the apply_action to be compliant with the new rclone wrapper library I have done. This might be hard to do...